feat(details): Add favorite button and feedback

+++ composeApp/src/commonMain/kotlin/zed/rainxch/githubstore/feature/details/presentation/DetailsEvent.kt
@@ -1,7 +1,7 @@
 package zed.rainxch.githubstore.feature.details.presentation
 
 sealed interface DetailsEvent {
-    data class OnOpenRepositoryInApp(val repositoryId: Int) : DetailsEvent
+    data class OnOpenRepositoryInApp(val repositoryId: Long) : DetailsEvent
     data class InstallTrackingFailed(val message: String) : DetailsEvent
     data class OnMessage(val message: String) : DetailsEvent
 }
